Moved Intents for Openstack to separate directory
